Fix warnings and incorrect usages of table headers
- fix usages of `UITableViewHeaderFooterView`
- fix warnings caused to missing items in Info.plist
