+++ UnstoppableWallet/UnstoppableWallet/AppIconAlternate.xcassets/AppIcon8ball.appiconset/Contents.json
@@ -52,17 +52,6 @@
       "idiom" : "ios-marketing",
       "scale" : "1x",
       "size" : "1024x1024"
-    },
-    {
-      "filename" : "settings58 1.png",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "filename" : "settings87 1.png",
-      "idiom" : "universal",
-      "scale" : "1x",
-      "unassigned" : true
     }
   ],
   "info" : {
