+++ UnstoppableWallet/UnstoppableWallet/Modules/NftAsset/NftAssetViewController.swift
@@ -7,7 +7,7 @@ import HUD
 import ComponentKit
 
 class NftAssetViewController: ThemeViewController {
-    private let tabsView = FilterHeaderView(buttonStyle: .tab)
+    private let tabsView = FilterView(buttonStyle: .tab)
     private let pageViewController = UIPageViewController(transitionStyle: .scroll, navigationOrientation: .horizontal)
 
     private let overviewController: NftAssetOverviewViewController
@@ -37,7 +37,7 @@ class NftAssetViewController: ThemeViewController {
         tabsView.snp.makeConstraints { maker in
             maker.top.equalTo(view.safeAreaLayoutGuide)
             maker.leading.trailing.equalToSuperview()
-            maker.height.equalTo(FilterHeaderView.height)
+            maker.height.equalTo(FilterView.height)
         }
 
         view.addSubview(pageViewController.view)
@@ -47,7 +47,7 @@ class NftAssetViewController: ThemeViewController {
         }
 
         tabsView.reload(filters: NftAssetModule.Tab.allCases.map {
-            FilterHeaderView.ViewItem.item(title: $0.title)
+            FilterView.ViewItem.item(title: $0.title)
         })
 
         tabsView.onSelect = { [weak self] index in
