+++ UnstoppableWallet/UnstoppableWallet/Modules/NftCollection/NftCollectionViewController.swift
@@ -11,7 +11,7 @@ class NftCollectionViewController: ThemeViewController {
     private let viewModel: NftCollectionViewModel
     private let disposeBag = DisposeBag()
 
-    private let tabsView = FilterHeaderView(buttonStyle: .tab)
+    private let tabsView = FilterView(buttonStyle: .tab)
     private let pageViewController = UIPageViewController(transitionStyle: .scroll, navigationOrientation: .horizontal)
 
     private let overviewController: NftCollectionOverviewViewController
@@ -44,7 +44,7 @@ class NftCollectionViewController: ThemeViewController {
         tabsView.snp.makeConstraints { maker in
             maker.top.equalTo(view.safeAreaLayoutGuide)
             maker.leading.trailing.equalToSuperview()
-            maker.height.equalTo(FilterHeaderView.height)
+            maker.height.equalTo(FilterView.height)
         }
 
         view.addSubview(pageViewController.view)
@@ -54,7 +54,7 @@ class NftCollectionViewController: ThemeViewController {
         }
 
         tabsView.reload(filters: NftCollectionModule.Tab.allCases.map {
-            FilterHeaderView.ViewItem.item(title: $0.title)
+            FilterView.ViewItem.item(title: $0.title)
         })
 
         tabsView.onSelect = { [weak self] index in
